feat: Add endpoint to retrieve a list of all books
Implemented a new endpoint, /books, that responds to HTTP GET requests and returns a list of all books. The method leverages the listBooks method from the BookService to fetch the data. If books are found, it responds with a list and HTTP status OK; otherwise, it returns HTTP status NOT_FOUND."